Add dependency in build.rs
