+++ Cargo.toml
@@ -234,6 +234,9 @@ panic_hook = { path = "components/panic_hook" }
 tokio = { version = "1.5", features = ["macros", "rt-multi-thread", "time"] }
 zipf = "6.1.0"
 
+[build-dependencies]
+gen_proxy_ffi = { path = "gen-proxy-ffi", default-features = false, package = "gen-proxy-ffi" }
+
 
 [patch.crates-io]
 prometheus = { git = "https://github.com/solotzg/rust-prometheus.git", rev = "3bc64acb5f7a23685f9c8f54d4067eb4273260c7" }
