+++ build.rs
@@ -1,0 +1,3 @@
+fn main() {
+    gen_proxy_ffi::gen_ffi_code();
+}
